+++ euler/src/solvers/p0005.rs
@@ -4,7 +4,7 @@ pub fn p0005_solver() -> String {
     smallest_multiple(20).to_string()
 }
 
-fn smallest_multiple(max: usize) -> usize {
+fn smallest_multiple(max: u64) -> u64 {
     (1..=max)
         .reduce(|acc, cur| lcm(acc, cur))
         .unwrap()
