+++ euler/src/solvers/p0006.rs
@@ -2,8 +2,8 @@ pub fn p0006_solver() -> String {
     sum_square_difference(100).to_string()
 }
 
-fn sum_square_difference(max: usize) -> usize {
-    usize::pow((1..=max).sum(), 2) - (1..=max).map(|x| x * x).sum::<usize>()
+fn sum_square_difference(max: u64) -> u64 {
+    u64::pow((1..=max).sum(), 2) - (1..=max).map(|x| x * x).sum::<u64>()
 }
 
 #[test]
