+++ euler/src/solvers/p0009.rs
@@ -2,7 +2,7 @@ pub fn p0009_solver() -> String {
     special_pythagorean_triplet(1000).to_string()
 }
 
-fn special_pythagorean_triplet(sum: usize) -> usize {
+fn special_pythagorean_triplet(sum: u64) -> u64 {
     (1..=sum)
         .flat_map(|a| (a+1..sum-a).map(move |b| (a, b)))
         .map(|(a,b)| (a,b,sum-a-b))
