+++ euler/src/solvers/p0010.rs
@@ -2,28 +2,28 @@ pub fn p0010_solver() -> String {
     summation_of_primes(2_000_000).to_string()
 }
 
-fn summation_of_primes(max: usize) -> usize {
+fn summation_of_primes(max: u64) -> u64 {
     let mut sieve: Vec<bool> = vec!();
-    sieve.resize(max, true);
+    sieve.resize(max as usize, true);
     sieve[0] = false;
     sieve[1] = false;
     let mut i = 1;
     while i < max - 1 {
         i += 1;
-        if !sieve[i] {
+        if !sieve[i as usize] {
             continue;
         }
 
         let mut j = i * 2;
         while j < max {
-            sieve[j] = false;
+            sieve[j as usize] = false;
             j += i;
         }
     }
 
     sieve.into_iter()
         .enumerate()
-        .filter_map(|(i,s)| if s { Some(i) } else { None })
+        .filter_map(|(i,s)| if s { Some(i as u64) } else { None })
         .sum()
 }
