+++ euler/src/solvers/p0014.rs
@@ -5,8 +5,8 @@ pub fn p0014_solver() -> String {
     longest_collatz_sequence(1_000_000).to_string()
 }
 
-fn longest_collatz_sequence(max: u128) -> u128 {
-    let mut map: HashMap<u128, u128> = HashMap::new();
+fn longest_collatz_sequence(max: u64) -> u64 {
+    let mut map: HashMap<u64, usize> = HashMap::new();
     map.insert(1, 1);
     for i in 2..max {
         calculate_collatz(i, &mut map);
@@ -19,7 +19,7 @@ fn longest_collatz_sequence(max: u128) -> u128 {
         .0
 }
 
-fn calculate_collatz(value: u128, map: &mut HashMap<u128, u128>) -> u128 {
+fn calculate_collatz(value: u64, map: &mut HashMap<u64, usize>) -> usize {
     if let Some(c) = map.get(&value) { return *c; }
     let n = calculate_collatz(if value % 2 == 0 { value / 2 } else { value * 3 + 1 }, map) + 1;
     map.insert(value, n);
