+++ euler/src/solvers/p0022.rs
@@ -5,22 +5,22 @@ pub fn p0022_solver() -> String {
     name_scores(&read_to_string("input/0022_names.txt").unwrap()).to_string()
 }
 
-fn name_scores(input: &str) -> u128 {
+fn name_scores(input: &str) -> u64 {
     input.split(',')
         .map(|n| n.replace('"', ""))
         .sorted()
         .enumerate()
-        .map(|(i,n)| name_score(&n) * (i as u128 + 1))
+        .map(|(i,n)| name_score(&n) * (i as u64 + 1))
         .sum()
 }
 
-fn name_score(name: &str) -> u128 {
+fn name_score(name: &str) -> u64 {
     name.chars()
-        .map(|c| c as u128 - LETTER_ZERO)
+        .map(|c| c as u64 - LETTER_ZERO)
         .sum()
 }
 
-static LETTER_ZERO: u128 = 'A' as u128 - 1;
+static LETTER_ZERO: u64 = 'A' as u64 - 1;
 
 #[test]
 fn computes_score_for_name() {
